Made tests for LogicalModels

+++ api/views/LogicalModels.py
@@ -2,6 +2,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.exceptions import PermissionDenied
+
 from django.http import Http404
 
 from api.models import LogicalModel
@@ -17,26 +18,28 @@ class LogicalModels(APIView):
 
 		models = LogicalModel.objects.filter(user=request.user)
 		serializer = LogicalModelSerializer(models, many=True)
+
 		return Response(serializer.data)
 
 
 	def post(self, request):
 
-		try:
-			model = LogicalModel(
-				user=request.user,
-				name=request.data['name'],
-				file=request.data['file']
-			)
+		if request.user.is_anonymous:
+			raise PermissionDenied
 
-			model.save()
+		LogicalModel(
+			user=request.user,
+			name=request.data['name'],
+			file=request.data['file']
+		).save()
 
-			return Response(status=status.HTTP_200_OK)
+		return Response(status=status.HTTP_200_OK)
 
-		except:
-			raise Http404
 
 	def delete(self, request, pk=None, format=None):
+
+		if request.user.is_anonymous:
+			raise PermissionDenied
 		try:
 			model = LogicalModel.objects.get(user=request.user, id=request.data['id'])
 			model.delete()
